+++ packages/article-paragraph/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.9.363](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.9.362...@times-components/article-paragraph@1.9.363) (2025-01-20)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
 ## [1.9.362](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.9.361...@times-components/article-paragraph@1.9.362) (2025-01-09)
 
 **Note:** Version bump only for package @times-components/article-paragraph