+++ packages/article-summary/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.25.29](https://github.com/newsuk/times-components/compare/@times-components/article-summary@3.25.28...@times-components/article-summary@3.25.29) (2025-01-20)
+
+**Note:** Version bump only for package @times-components/article-summary
+
+
+
+
+
 ## [3.25.28](https://github.com/newsuk/times-components/compare/@times-components/article-summary@3.25.27...@times-components/article-summary@3.25.28) (2025-01-09)
 
 **Note:** Version bump only for package @times-components/article-summary