+++ packages/author-profile/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.19.21](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.19.20...@times-components/author-profile@6.19.21) (2025-01-20)
+
+**Note:** Version bump only for package @times-components/author-profile
+
+
+
+
+
 ## [6.19.20](https://github.com/newsuk/times-components/compare/@times-components/author-profile@6.19.19...@times-components/author-profile@6.19.20) (2025-01-09)
 
 **Note:** Version bump only for package @times-components/author-profile